Wait for map mixin
Add single shared mixin with map vm waiter.
Cancel map vm waiter also on current vm licycle cancel.
