+++ src/mixins/base-layer.js
@@ -1,17 +1,14 @@
 import debounce from 'debounce-promise'
 import { get as getProj } from 'ol/proj'
 import { from as fromObs, merge as mergeObs } from 'rxjs'
-import { filter as filterObs, map as mapObs, mapTo, mergeMap, skipWhile } from 'rxjs/operators'
+import { map as mapObs, mergeMap, skipWhile } from 'rxjs/operators'
 import { EPSG_3857, getLayerId, initializeLayer, roundExtent, setLayerId, transformExtent } from '../ol-ext'
-import {
-  fromOlChangeEvent as obsFromOlChangeEvent,
-  fromVueEvent as obsFromVueEvent,
-  fromVueWatcher as obsFromVueWatcher,
-} from '../rx-ext'
-import { addPrefix, assert, hasProp, isEqual, isNumber, mergeDescriptors, pick, stubTrue, waitFor } from '../utils'
-import olCmp, { FRAME_TIME, isCreateError, OlObjectEvent } from './ol-cmp'
+import { fromOlChangeEvent as obsFromOlChangeEvent } from '../rx-ext'
+import { addPrefix, assert, isEqual, isNumber, mergeDescriptors, pick } from '../utils'
+import olCmp, { FRAME_TIME } from './ol-cmp'
 import projTransforms from './proj-transforms'
 import stubVNode from './stub-vnode'
+import waitForMap from './wait-for-map'
 
 /**
  * Base layer mixin.
@@ -21,6 +18,7 @@ export default {
     stubVNode,
     projTransforms,
     olCmp,
+    waitForMap,
   ],
   stubVNode: {
     attrs () {
@@ -237,41 +235,6 @@ export default {
   },
   methods: {
     /**
-     * @returns {Promise<void>}
-     * @protected
-     */
-    async beforeInit () {
-      try {
-        await waitFor(
-          () => this.$mapVm != null,
-          obsFromVueEvent(this.$eventBus, OlObjectEvent.ERROR).pipe(
-            filterObs(([err, vm]) => {
-              return isCreateError(err) &&
-                hasProp(vm, '$map') &&
-                this.$vq.closest(vm)
-            }),
-            mapTo(stubTrue()),
-          ),
-        )
-        this.viewProjection = this.$mapVm.resolvedViewProjection
-        this.dataProjection = this.$mapVm.resolvedDataProjection
-        this.subscribeTo(
-          obsFromVueWatcher(this, () => this.$mapVm.resolvedViewProjection),
-          ({ value }) => { this.viewProjection = value },
-        )
-        this.subscribeTo(
-          obsFromVueWatcher(this, () => this.$mapVm.resolvedDataProjection),
-          ({ value }) => { this.dataProjection = value },
-        )
-        await this.$nextTickPromise()
-
-        return this::olCmp.methods.beforeInit()
-      } catch (err) {
-        err.message = `${this.vmName} wait for $mapVm injection: ${err.message}`
-        throw err
-      }
-    },
-    /**
      * @return {Promise<module:ol/layer/Base~BaseLayer>}
      * @protected
      */
@@ -345,6 +308,9 @@ export default {
       'recreate',
       'scheduleRecreate',
       'resolveOlObject',
+    ]),
+    .../*#__PURE__*/pick(waitForMap.methods, [
+      'beforeInit',
     ]),
     /**
      * @returns {string|number}
