+++ src/mixins/geometry.js
@@ -1,14 +1,11 @@
-import { filter as filterObs, mapTo, skipWhile } from 'rxjs/operators'
+import { skipWhile } from 'rxjs/operators'
 import { EPSG_3857, getGeometryId, initializeGeometry, roundExtent, roundPointCoords, setGeometryId } from '../ol-ext'
-import {
-  fromOlChangeEvent as obsFromOlChangeEvent,
-  fromVueEvent as obsFromVueEvent,
-  fromVueWatcher as obsFromVueWatcher,
-} from '../rx-ext'
-import { addPrefix, assert, hasProp, isEqual, mergeDescriptors, pick, stubTrue, waitFor } from '../utils'
-import olCmp, { isCreateError, OlObjectEvent } from './ol-cmp'
+import { fromOlChangeEvent as obsFromOlChangeEvent } from '../rx-ext'
+import { addPrefix, assert, isEqual, mergeDescriptors, pick } from '../utils'
+import olCmp from './ol-cmp'
 import projTransforms from './proj-transforms'
 import stubVNode from './stub-vnode'
+import waitForMap from './wait-for-map'
 
 /**
  * Base geometry mixin.
@@ -18,6 +15,7 @@ export default {
     stubVNode,
     projTransforms,
     olCmp,
+    waitForMap,
   ],
   stubVNode: {
     empty () {
@@ -51,41 +49,6 @@ export default {
     this::defineServices()
   },
   methods: {
-    /**
-     * @returns {Promise<void>}
-     * @protected
-     */
-    async beforeInit () {
-      try {
-        await waitFor(
-          () => this.$mapVm != null,
-          obsFromVueEvent(this.$eventBus, OlObjectEvent.ERROR).pipe(
-            filterObs(([err, vm]) => {
-              return isCreateError(err) &&
-                hasProp(vm, '$map') &&
-                this.$vq.closest(vm)
-            }),
-            mapTo(stubTrue()),
-          ),
-        )
-        this.viewProjection = this.$mapVm.resolvedViewProjection
-        this.dataProjection = this.$mapVm.resolvedDataProjection
-        this.subscribeTo(
-          obsFromVueWatcher(this, () => this.$mapVm.resolvedViewProjection),
-          ({ value }) => { this.viewProjection = value },
-        )
-        this.subscribeTo(
-          obsFromVueWatcher(this, () => this.$mapVm.resolvedDataProjection),
-          ({ value }) => { this.dataProjection = value },
-        )
-        await this.$nextTickPromise()
-
-        return this::olCmp.methods.beforeInit()
-      } catch (err) {
-        err.message = `${this.vmName} wait for $mapVm injection: ${err.message}`
-        throw err
-      }
-    },
     /**
      * @return {Promise<module:ol/geom/Geometry~Geometry>}
      * @protected
@@ -159,6 +122,9 @@ export default {
       'recreate',
       'scheduleRecreate',
       'resolveOlObject',
+    ]),
+    .../*#__PURE__*/pick(waitForMap.methods, [
+      'beforeInit',
     ]),
     /**
      * @returns {string|number}
