+++ src/mixins/index.js
@@ -48,5 +48,6 @@ export tileSource from './tile-source'
 export urlTileSource from './url-tile-source'
 export vectorLayer from './vector-layer'
 export vectorSource from './vector-source'
+export waitForMap from './wait-for-map'
 export wmsSource from './wms-source'
 export xyzSource from './xyz-source'
