+++ src/mixins/interaction.js
@@ -1,5 +1,5 @@
 import debounce from 'debounce-promise'
-import { filter as filterObs, mapTo, skipWhile } from 'rxjs/operators'
+import { skipWhile } from 'rxjs/operators'
 import {
   EPSG_3857,
   getInteractionId,
@@ -8,15 +8,12 @@ import {
   setInteractionId,
   setInteractionPriority,
 } from '../ol-ext'
-import {
-  fromOlChangeEvent as obsFromOlChangeEvent,
-  fromVueEvent as obsFromVueEvent,
-  fromVueWatcher as obsFromVueWatcher,
-} from '../rx-ext'
-import { addPrefix, assert, hasProp, isEqual, mergeDescriptors, pick, stubTrue, waitFor } from '../utils'
-import olCmp, { FRAME_TIME, isCreateError, OlObjectEvent } from './ol-cmp'
+import { fromOlChangeEvent as obsFromOlChangeEvent } from '../rx-ext'
+import { addPrefix, assert, isEqual, mergeDescriptors, pick } from '../utils'
+import olCmp, { FRAME_TIME } from './ol-cmp'
 import projTransforms from './proj-transforms'
 import stubVNode from './stub-vnode'
+import waitForMap from './wait-for-map'
 
 /**
  * Base interaction mixin.
@@ -26,6 +23,7 @@ export default {
     stubVNode,
     projTransforms,
     olCmp,
+    waitForMap,
   ],
   stubVNode: {
     empty () {
@@ -94,41 +92,6 @@ export default {
     this::defineServices()
   },
   methods: {
-    /**
-     * @returns {Promise<void>}
-     * @protected
-     */
-    async beforeInit () {
-      try {
-        await waitFor(
-          () => this.$mapVm != null,
-          obsFromVueEvent(this.$eventBus, OlObjectEvent.ERROR).pipe(
-            filterObs(([err, vm]) => {
-              return isCreateError(err) &&
-                hasProp(vm, '$map') &&
-                this.$vq.closest(vm)
-            }),
-            mapTo(stubTrue()),
-          ),
-        )
-        this.viewProjection = this.$mapVm.resolvedViewProjection
-        this.dataProjection = this.$mapVm.resolvedDataProjection
-        this.subscribeTo(
-          obsFromVueWatcher(this, () => this.$mapVm.resolvedViewProjection),
-          ({ value }) => { this.viewProjection = value },
-        )
-        this.subscribeTo(
-          obsFromVueWatcher(this, () => this.$mapVm.resolvedDataProjection),
-          ({ value }) => { this.dataProjection = value },
-        )
-        await this.$nextTickPromise()
-
-        return this::olCmp.methods.beforeInit()
-      } catch (err) {
-        err.message = `${this.vmName} wait for $mapVm injection: ${err.message}`
-        throw err
-      }
-    },
     /**
      * @return {Promise<module:ol/interaction/Interaction~Interaction>}
      * @protected
@@ -209,6 +172,9 @@ export default {
       'remount',
       'scheduleRemount',
       'resolveOlObject',
+    ]),
+    .../*#__PURE__*/pick(waitForMap.methods, [
+      'beforeInit',
     ]),
     /**
      * @returns {string|number}
