+++ src/mixins/source.js
@@ -1,27 +1,13 @@
 import debounce from 'debounce-promise'
 import { get as getProj } from 'ol/proj'
-import { filter as filterObs, mapTo, skipWhile } from 'rxjs/operators'
+import { skipWhile } from 'rxjs/operators'
 import { EPSG_3857, getSourceId, initializeSource, setSourceId } from '../ol-ext'
-import {
-  fromOlChangeEvent as obsFromOlChangeEvent,
-  fromVueEvent as obsFromVueEvent,
-  fromVueWatcher as obsFromVueWatcher,
-} from '../rx-ext'
-import {
-  addPrefix,
-  assert,
-  hasProp,
-  isArray,
-  isEqual,
-  isString,
-  mergeDescriptors,
-  pick,
-  stubTrue,
-  waitFor,
-} from '../utils'
-import olCmp, { FRAME_TIME, isCreateError, OlObjectEvent } from './ol-cmp'
+import { fromOlChangeEvent as obsFromOlChangeEvent } from '../rx-ext'
+import { addPrefix, assert, isArray, isEqual, isString, mergeDescriptors, pick } from '../utils'
+import olCmp, { FRAME_TIME } from './ol-cmp'
 import projTransforms from './proj-transforms'
 import stubVNode from './stub-vnode'
+import waitForMap from './wait-for-map'
 
 /**
  * Base source mixin.
@@ -31,6 +17,7 @@ export default {
     stubVNode,
     projTransforms,
     olCmp,
+    waitForMap,
   ],
   stubVNode: {
     empty () {
@@ -143,41 +130,6 @@ export default {
   },
   methods: {
     /**
-     * @returns {Promise<void>}
-     * @protected
-     */
-    async beforeInit () {
-      try {
-        await waitFor(
-          () => this.$mapVm != null,
-          obsFromVueEvent(this.$eventBus, OlObjectEvent.ERROR).pipe(
-            filterObs(([err, vm]) => {
-              return isCreateError(err) &&
-                hasProp(vm, '$map') &&
-                this.$vq.closest(vm)
-            }),
-            mapTo(stubTrue()),
-          ),
-        )
-        this.viewProjection = this.$mapVm.resolvedViewProjection
-        this.dataProjection = this.$mapVm.resolvedDataProjection
-        this.subscribeTo(
-          obsFromVueWatcher(this, () => this.$mapVm.resolvedViewProjection),
-          ({ value }) => { this.viewProjection = value },
-        )
-        this.subscribeTo(
-          obsFromVueWatcher(this, () => this.$mapVm.resolvedDataProjection),
-          ({ value }) => { this.dataProjection = value },
-        )
-        await this.$nextTickPromise()
-
-        return this::olCmp.methods.beforeInit()
-      } catch (err) {
-        err.message = `${this.vmName} wait for $mapVm injection: ${err.message}`
-        throw err
-      }
-    },
-    /**
      * @return {Promise<module:ol/source/Source~Source>}
      * @protected
      */
@@ -257,6 +209,9 @@ export default {
       'recreate',
       'scheduleRecreate',
       'resolveOlObject',
+    ]),
+    .../*#__PURE__*/pick(waitForMap.methods, [
+      'beforeInit',
     ]),
     /**
      * @return {Promise<module:ol/source/Source~Source>}
